fix: 🐛 address comments in PR
