+++ src/components/DataRoom/Hero/utils/asciiShader.tsx
@@ -1,115 +1,5 @@
 import { useEffect, useMemo } from 'react'
-import { Effect } from 'postprocessing'
-import { CanvasTexture, Color, NearestFilter, RepeatWrapping, Texture, Uniform } from 'three'
-
-type ASCIIEffectOptions = {
-  characters?: string
-  fontSize?: number
-  cellSize?: number
-  color?: string
-  invert?: boolean
-}
-
-const SIZE = 1024
-const MAX_PER_ROW = 16
-const CELL = SIZE / MAX_PER_ROW
-
-// ASCII Effect Fragment Shader
-// This shader converts the input image to ASCII art by mapping pixel intensities to characters
-// and applying color, inversion, and pixelization effects.
-const fragmentShader = `
-uniform sampler2D uCharacters;
-uniform float uCharactersCount;
-uniform float uCellSize;
-uniform bool uInvert;
-uniform vec3 uColor;
-
-const vec2 SIZE = vec2(16.);
-
-vec3 greyscale(vec3 color, float strength) {
-    float g = dot(color, vec3(0.299, 0.587, 0.114));
-    return mix(color, vec3(g), strength);
-}
-
-vec3 greyscale(vec3 color) {
-    return greyscale(color, 1.0);
-}
-
-void mainImage(const in vec4 inputColor, const in vec2 uv, out vec4 outputColor) {
-    vec2 cell = resolution / uCellSize;
-    vec2 grid = 1.0 / cell;
-    vec2 pixelizedUV = grid * (0.5 + floor(uv / grid));
-    vec4 pixelized = texture2D(inputBuffer, pixelizedUV);
-    float greyscaled = greyscale(pixelized.rgb).r;
-
-    if (uInvert) {
-        greyscaled = 1.0 - greyscaled;
-    }
-
-    float characterIndex = floor((uCharactersCount - 1.0) * greyscaled);
-    vec2 characterPosition = vec2(mod(characterIndex, SIZE.x), floor(characterIndex / SIZE.y));
-    vec2 offset = vec2(characterPosition.x, -characterPosition.y) / SIZE;
-    vec2 charUV = mod(uv * (cell / SIZE), 1.0 / SIZE) - vec2(0., 1.0 / SIZE) + offset;
-    vec4 asciiCharacter = texture2D(uCharacters, charUV);
-
-    asciiCharacter.rgb = uColor * asciiCharacter.r;
-    asciiCharacter.a = pixelized.a;
-    outputColor = asciiCharacter;
-}
-`
-
-// ASCIIEffect extends the Effect class from the postprocessing library.
-// Effect is a base class for custom postprocessing effects in Three.js.
-// By extending it, we create a specialized effect for rendering ASCII art.
-export class ASCIIEffect extends Effect {
-  constructor({
-    characters = ` .:,'-^=*+?!|0#X%WM@`,
-    fontSize = 64,
-    cellSize = 28,
-    color = '#12FF80',
-    invert = true,
-  }: ASCIIEffectOptions = {}) {
-    const uniforms = new Map<string, Uniform<any>>([
-      ['uCharacters', new Uniform(new Texture())],
-      ['uCellSize', new Uniform(cellSize)],
-      ['uCharactersCount', new Uniform(characters.length)],
-      ['uColor', new Uniform(new Color(color))],
-      ['uInvert', new Uniform(invert)],
-    ])
-
-    super('ASCIIEffect', fragmentShader, { uniforms })
-
-    this.createCharactersTexture(characters, fontSize)
-  }
-
-  private createCharactersTexture(characters: string, fontSize: number): void {
-    const canvas = document.createElement('canvas')
-    canvas.width = canvas.height = SIZE
-
-    const context = canvas.getContext('2d')
-    if (!context) throw new Error('Context not available')
-
-    context.font = `${fontSize}px Citerne`
-    context.textAlign = 'center'
-    context.textBaseline = 'middle'
-    context.fillStyle = '#FFF'
-
-    characters.split('').forEach((char, i) => {
-      const x = (i % MAX_PER_ROW) * CELL + CELL / 2
-      const y = Math.floor(i / MAX_PER_ROW) * CELL + CELL / 2
-      context.fillText(char, x, y)
-    })
-
-    const texture = new CanvasTexture(canvas)
-    texture.minFilter = NearestFilter
-    texture.magFilter = NearestFilter
-    texture.wrapS = texture.wrapT = RepeatWrapping
-    texture.needsUpdate = true
-
-    const charactersTextureUniform = this.uniforms.get('uCharacters')
-    if (charactersTextureUniform) charactersTextureUniform.value = texture
-  }
-}
+import { ASCIIEffect } from './asciiEffect'
 
 // This component wraps and manages the lifecycle of the ASCII effect,
 // It creates the effect, handles its disposal, and renders it as a primitive object.
